requirements.txt
langchain
langchain-community
langchain-chroma
langchain-google-genai
langchain-groq
langchain-huggingface
pymupdf
chromadb
sentence-transformers
python-dotenv
